+++ lib/money/money/formatter.rb
@@ -13,8 +13,8 @@ class Money
     #  amount of money should be formatted of "free" or as the supplied string.
     #
     # @example
-    #   Money.us_dollar(0).format(:display_free => true)     #=> "free"
-    #   Money.us_dollar(0).format(:display_free => "gratis") #=> "gratis"
+    #   Money.us_dollar(0).format(display_free: true)     #=> "free"
+    #   Money.us_dollar(0).format(display_free: "gratis") #=> "gratis"
     #   Money.us_dollar(0).format                            #=> "$0.00"
     #
     # @option rules [Boolean] :with_currency (false) Whether the currency name
@@ -22,29 +22,29 @@ class Money
     #
     # @example
     #   Money.ca_dollar(100).format #=> "$1.00"
-    #   Money.ca_dollar(100).format(:with_currency => true) #=> "$1.00 CAD"
-    #   Money.us_dollar(85).format(:with_currency => true)  #=> "$0.85 USD"
+    #   Money.ca_dollar(100).format(with_currency: true) #=> "$1.00 CAD"
+    #   Money.us_dollar(85).format(with_currency: true)  #=> "$0.85 USD"
     #
     # @option rules [Boolean] :rounded_infinite_precision (false) Whether the
     #  amount of money should be rounded when using {infinite_precision}
     #
     # @example
     #   Money.us_dollar(100.1).format #=> "$1.001"
-    #   Money.us_dollar(100.1).format(:rounded_infinite_precision => true) #=> "$1"
-    #   Money.us_dollar(100.9).format(:rounded_infinite_precision => true) #=> "$1.01"
+    #   Money.us_dollar(100.1).format(rounded_infinite_precision: true) #=> "$1"
+    #   Money.us_dollar(100.9).format(rounded_infinite_precision: true) #=> "$1.01"
     #
     # @option rules [Boolean] :no_cents (false) Whether cents should be omitted.
     #
     # @example
-    #   Money.ca_dollar(100).format(:no_cents => true) #=> "$1"
-    #   Money.ca_dollar(599).format(:no_cents => true) #=> "$5"
+    #   Money.ca_dollar(100).format(no_cents: true) #=> "$1"
+    #   Money.ca_dollar(599).format(no_cents: true) #=> "$5"
     #
     # @option rules [Boolean] :no_cents_if_whole (false) Whether cents should be
     #  omitted if the cent value is zero
     #
     # @example
-    #   Money.ca_dollar(10000).format(:no_cents_if_whole => true) #=> "$100"
-    #   Money.ca_dollar(10034).format(:no_cents_if_whole => true) #=> "$100.34"
+    #   Money.ca_dollar(10000).format(no_cents_if_whole: true) #=> "$100"
+    #   Money.ca_dollar(10034).format(no_cents_if_whole: true) #=> "$100.34"
     #
     # @option rules [Boolean, String, nil] :symbol (true) Whether a money symbol
     #  should be prepended to the result string. The default is true. This method
@@ -56,26 +56,26 @@ class Money
     #   Money.new(100, "EUR") #=> "€1.00"
     #
     #   # Same thing.
-    #   Money.new(100, "USD").format(:symbol => true) #=> "$1.00"
-    #   Money.new(100, "GBP").format(:symbol => true) #=> "£1.00"
-    #   Money.new(100, "EUR").format(:symbol => true) #=> "€1.00"
+    #   Money.new(100, "USD").format(symbol: true) #=> "$1.00"
+    #   Money.new(100, "GBP").format(symbol: true) #=> "£1.00"
+    #   Money.new(100, "EUR").format(symbol: true) #=> "€1.00"
     #
     #   # You can specify a false expression or an empty string to disable
     #   # prepending a money symbol.§
-    #   Money.new(100, "USD").format(:symbol => false) #=> "1.00"
-    #   Money.new(100, "GBP").format(:symbol => nil)   #=> "1.00"
-    #   Money.new(100, "EUR").format(:symbol => "")    #=> "1.00"
+    #   Money.new(100, "USD").format(symbol: false) #=> "1.00"
+    #   Money.new(100, "GBP").format(symbol: nil)   #=> "1.00"
+    #   Money.new(100, "EUR").format(symbol: "")    #=> "1.00"
     #
     #   # If the symbol for the given currency isn't known, then it will default
     #   # to "¤" as symbol.
-    #   Money.new(100, "AWG").format(:symbol => true) #=> "¤1.00"
+    #   Money.new(100, "AWG").format(symbol: true) #=> "¤1.00"
     #
     #   # You can specify a string as value to enforce using a particular symbol.
-    #   Money.new(100, "AWG").format(:symbol => "ƒ") #=> "ƒ1.00"
+    #   Money.new(100, "AWG").format(symbol: "ƒ") #=> "ƒ1.00"
     #
     #   # You can specify a indian currency format
-    #   Money.new(10000000, "INR").format(:south_asian_number_formatting => true) #=> "1,00,000.00"
-    #   Money.new(10000000).format(:south_asian_number_formatting => true) #=> "$1,00,000.00"
+    #   Money.new(10000000, "INR").format(south_asian_number_formatting: true) #=> "1,00,000.00"
+    #   Money.new(10000000).format(south_asian_number_formatting: true) #=> "$1,00,000.00"
     #
     # @option rules [Boolean, nil] :symbol_before_without_space (true) Whether
     #   a space between the money symbol and the amount should be inserted when
@@ -87,10 +87,10 @@ class Money
     #   Money.new(100, "USD").format #=> "$1.00"
     #
     #   # Same thing.
-    #   Money.new(100, "USD").format(:symbol_before_without_space => true) #=> "$1.00"
+    #   Money.new(100, "USD").format(symbol_before_without_space: true) #=> "$1.00"
     #
     #   # If set to false, will insert a space.
-    #   Money.new(100, "USD").format(:symbol_before_without_space => false) #=> "$ 1.00"
+    #   Money.new(100, "USD").format(symbol_before_without_space: false) #=> "$ 1.00"
     #
     # @option rules [Boolean, nil] :symbol_after_without_space (false) Whether
     #   a space between the amount and the money symbol should be inserted when
@@ -99,17 +99,17 @@ class Money
     #
     # @example
     #   # Default is to insert a space.
-    #   Money.new(100, "USD").format(:symbol_position => :after) #=> "1.00 $"
+    #   Money.new(100, "USD").format(symbol_position: :after) #=> "1.00 $"
     #
     #   # If set to true, will not insert a space.
-    #   Money.new(100, "USD").format(:symbol_position => :after, :symbol_after_without_space => true) #=> "1.00$"
+    #   Money.new(100, "USD").format(symbol_position: :after, symbol_after_without_space: true) #=> "1.00$"
     #
     # @option rules [Boolean, String, nil] :decimal_mark (true) Whether the
     #  currency should be separated by the specified character or '.'
     #
     # @example
     #   # If a string is specified, it's value is used.
-    #   Money.new(100, "USD").format(:decimal_mark => ",") #=> "$1,00"
+    #   Money.new(100, "USD").format(decimal_mark: ",") #=> "$1,00"
     #
     #   # If the decimal_mark for a given currency isn't known, then it will default
     #   # to "." as decimal_mark.
@@ -120,12 +120,12 @@ class Money
     #
     # @example
     #   # If false is specified, no thousands_separator is used.
-    #   Money.new(100000, "USD").format(:thousands_separator => false) #=> "1000.00"
-    #   Money.new(100000, "USD").format(:thousands_separator => nil)   #=> "1000.00"
-    #   Money.new(100000, "USD").format(:thousands_separator => "")    #=> "1000.00"
+    #   Money.new(100000, "USD").format(thousands_separator: false) #=> "1000.00"
+    #   Money.new(100000, "USD").format(thousands_separator: nil)   #=> "1000.00"
+    #   Money.new(100000, "USD").format(thousands_separator: "")    #=> "1000.00"
     #
     #   # If a string is specified, it's value is used.
-    #   Money.new(100000, "USD").format(:thousands_separator => ".") #=> "$1.000.00"
+    #   Money.new(100000, "USD").format(thousands_separator: ".") #=> "$1.000.00"
     #
     #   # If the thousands_separator for a given currency isn't known, then it will
     #   # default to "," as thousands_separator.
@@ -135,16 +135,22 @@ class Money
     #  HTML-formatted. Only useful in combination with +:with_currency+.
     #
     # @example
-    #   Money.ca_dollar(570).format(:html => true, :with_currency => true)
+    #   Money.ca_dollar(570).format(html: true, with_currency: true)
     #   #=> "$5.70 <span class=\"currency\">CAD</span>"
+    #
+    # @option rules [Boolean] :html_wrap (false) Whether all currency parts should be HTML-formatted.
+    #
+    # @example
+    #   Money.ca_dollar(570).format(html_wrap: true, with_currency: true)
+    #   #=> "<span class=\"money-currency-symbol\">$</span><span class=\"money-whole\">5</span><span class=\"money-decimal-mark\">.</span><span class=\"money-decimal\">70</span> <span class=\"money-currency\">CAD</span>"
     #
     # @option rules [Boolean] :sign_before_symbol (false) Whether the sign should be
     #  before the currency symbol.
     #
     # @example
     #   # You can specify to display the sign before the symbol for negative numbers
-    #   Money.new(-100, "GBP").format(:sign_before_symbol => true)  #=> "-£1.00"
-    #   Money.new(-100, "GBP").format(:sign_before_symbol => false) #=> "£-1.00"
+    #   Money.new(-100, "GBP").format(sign_before_symbol: true)  #=> "-£1.00"
+    #   Money.new(-100, "GBP").format(sign_before_symbol: false) #=> "£-1.00"
     #   Money.new(-100, "GBP").format                               #=> "£-1.00"
     #
     # @option rules [Boolean] :sign_positive (false) Whether positive numbers should be
@@ -152,34 +158,34 @@ class Money
     #
     # @example
     #   # You can specify to display the sign with positive numbers
-    #   Money.new(100, "GBP").format(:sign_positive => true,  :sign_before_symbol => true)  #=> "+£1.00"
-    #   Money.new(100, "GBP").format(:sign_positive => true,  :sign_before_symbol => false) #=> "£+1.00"
-    #   Money.new(100, "GBP").format(:sign_positive => false, :sign_before_symbol => true)  #=> "£1.00"
-    #   Money.new(100, "GBP").format(:sign_positive => false, :sign_before_symbol => false) #=> "£1.00"
+    #   Money.new(100, "GBP").format(sign_positive: true,  sign_before_symbol: true)  #=> "+£1.00"
+    #   Money.new(100, "GBP").format(sign_positive: true,  sign_before_symbol: false) #=> "£+1.00"
+    #   Money.new(100, "GBP").format(sign_positive: false, sign_before_symbol: true)  #=> "£1.00"
+    #   Money.new(100, "GBP").format(sign_positive: false, sign_before_symbol: false) #=> "£1.00"
     #   Money.new(100, "GBP").format                               #=> "£+1.00"
     #
     # @option rules [Boolean] :disambiguate (false) Prevents the result from being ambiguous
     #  due to equal symbols for different currencies. Uses the `disambiguate_symbol`.
     #
     # @example
-    #   Money.new(10000, "USD").format(:disambiguate => false)   #=> "$100.00"
-    #   Money.new(10000, "CAD").format(:disambiguate => false)   #=> "$100.00"
-    #   Money.new(10000, "USD").format(:disambiguate => true)    #=> "$100.00"
-    #   Money.new(10000, "CAD").format(:disambiguate => true)    #=> "C$100.00"
+    #   Money.new(10000, "USD").format(disambiguate: false)   #=> "$100.00"
+    #   Money.new(10000, "CAD").format(disambiguate: false)   #=> "$100.00"
+    #   Money.new(10000, "USD").format(disambiguate: true)    #=> "$100.00"
+    #   Money.new(10000, "CAD").format(disambiguate: true)    #=> "C$100.00"
     #
     # @option rules [Boolean] :html_wrap_symbol (false) Wraps the currency symbol
     #  in a html <span> tag.
     #
     # @example
-    #   Money.new(10000, "USD").format(:disambiguate => false)
+    #   Money.new(10000, "USD").format(disambiguate: false)
     #   #=> "<span class=\"currency_symbol\">$100.00</span>
     #
     # @option rules [Symbol] :symbol_position (:before) `:before` if the currency
     #   symbol goes before the amount, `:after` if it goes after.
     #
     # @example
-    #   Money.new(10000, "USD").format(:symbol_position => :before) #=> "$100.00"
-    #   Money.new(10000, "USD").format(:symbol_position => :after)  #=> "100.00 $"
+    #   Money.new(10000, "USD").format(symbol_position: :before) #=> "$100.00"
+    #   Money.new(10000, "USD").format(symbol_position: :after)  #=> "100.00 $"
     #
     # @option rules [Boolean] :translate (true) `true` Checks for custom
     #   symbol definitions using I18n.
@@ -191,11 +197,18 @@ class Money
     #   #     currency:
     #   #       symbol:
     #   #         CAD: "CAD$"
-    #   Money.new(10000, "CAD").format(:translate => true) #=> "CAD$100.00"
+    #   Money.new(10000, "CAD").format(translate: true) #=> "CAD$100.00"
     #
     # @example
-    #   Money.new(89000, :btc).format(:drop_trailing_zeros => true) #=> B⃦0.00089
-    #   Money.new(110, :usd).format(:drop_trailing_zeros => true)   #=> $1.1
+    #   Money.new(89000, :btc).format(drop_trailing_zeros: true) #=> B⃦0.00089
+    #   Money.new(110, :usd).format(drop_trailing_zeros: true)   #=> $1.1
+    #
+    # @option rules [String] :format (nil) Provide a template for formatting. `%u` will be replaced
+    # with the symbol (if present) and `%n` will be replaced with the number.
+    #
+    # @example
+    #   Money.new(10000, "USD").format(format: '%u %n') #=> "$ 100.00"
+    #   Money.new(10000, "USD").format(format: '<span>%u%n</span>')  #=> "<span>$100.00</span>"
     #
     # Note that the default rules can be defined through {Money.default_formatting_rules} hash.
     #
@@ -209,6 +222,14 @@ class Money
     def to_s
       return free_text if show_free_text?
 
+      if rules[:html]
+        warn "[DEPRECATION] `html` is deprecated - use `html_wrap` instead. Please note that `html_wrap` will wrap all parts of currency and if you use `with_currency` option, currency element class changes from `currency` to `money-currency`."
+      end
+
+      if rules[:html_wrap_symbol]
+        warn "[DEPRECATION] `html_wrap_symbol` is deprecated - use `html_wrap` instead. Please note that `html_wrap` will wrap all parts of currency."
+      end
+
       whole_part, decimal_part = extract_whole_and_decimal_parts
 
       # Format whole and decimal parts separately
@@ -216,7 +237,19 @@ class Money
       whole_part = format_whole_part(whole_part)
 
       # Assemble the final formatted amount
-      formatted = [whole_part, decimal_part].compact.join(decimal_mark)
+      formatted = if rules[:html_wrap]
+        if decimal_part.nil?
+          html_wrap(whole_part, "whole")
+        else
+          [
+            html_wrap(whole_part, "whole"),
+            html_wrap(decimal_mark, "decimal-mark"),
+            html_wrap(decimal_part, "decimal")
+          ].join
+        end
+      else
+        [whole_part, decimal_part].compact.join(decimal_mark)
+      end
 
       sign = money.negative? ? '-' : ''
 
@@ -232,25 +265,29 @@ class Money
       symbol_value = symbol_value_from(rules)
 
       if symbol_value && !symbol_value.empty?
-        symbol_value = "<span class=\"currency_symbol\">#{symbol_value}</span>" if rules[:html_wrap_symbol]
-        symbol_position = symbol_position_from(rules)
-
-        formatted = if symbol_position == :before
-          symbol_space = rules[:symbol_before_without_space] === false ? " " : ""
-          "#{sign_before}#{symbol_value}#{symbol_space}#{sign}#{formatted}"
-        else
-          symbol_space = rules[:symbol_after_without_space] ? "" : " "
-          "#{sign_before}#{sign}#{formatted}#{symbol_space}#{symbol_value}"
+        if rules[:html_wrap_symbol]
+          symbol_value = "<span class=\"currency_symbol\">#{symbol_value}</span>"
+        elsif rules[:html_wrap]
+          symbol_value = html_wrap(symbol_value, "currency-symbol")
         end
+
+        formatted = rules[:format]
+                      .gsub('%u', [sign_before, symbol_value].join)
+                      .gsub('%n', [sign, formatted].join)
       else
-        formatted="#{sign_before}#{sign}#{formatted}"
+        formatted = "#{sign_before}#{sign}#{formatted}"
       end
 
       if rules[:with_currency]
         formatted << " "
-        formatted << '<span class="currency">' if rules[:html]
-        formatted << currency.to_s
-        formatted << '</span>' if rules[:html]
+
+        if rules[:html]
+          formatted << "<span class=\"currency\">#{currency.to_s}</span>"
+        elsif rules[:html_wrap]
+          formatted << html_wrap(currency.to_s, "currency")
+        else
+          formatted << currency.to_s
+        end
       end
       formatted
     end
@@ -282,20 +319,17 @@ class Money
       money.zero? && rules[:display_free]
     end
 
+    def html_wrap(string, class_name)
+      "<span class=\"money-#{class_name}\">#{string}</span>"
+    end
+
     def free_text
       rules[:display_free].respond_to?(:to_str) ? rules[:display_free] : 'free'
     end
 
     def format_whole_part(value)
-      # Determine thousands_separator
-      thousands_separator_value = if rules.has_key?(:thousands_separator)
-                                    rules[:thousands_separator] || ''
-                                  else
-                                    thousands_separator
-                                  end
-
       # Apply thousands_separator
-      value.gsub regexp_format, "\\1#{thousands_separator_value}"
+      value.gsub regexp_format, "\\1#{thousands_separator}"
     end
 
     def extract_whole_and_decimal_parts
@@ -328,9 +362,9 @@ class Money
     def i18n_format_for(method, name, character)
       if Money.use_i18n
         begin
-          I18n.t name, :scope => "number.currency.format", :raise => true
+          I18n.t name, scope: "number.currency.format", raise: true
         rescue I18n::MissingTranslationData
-          I18n.t name, :scope =>"number.format", :default => (currency.send(method) || character)
+          I18n.t name, scope: "number.format", default: (currency.send(method) || character)
         end
       else
         currency.send(method) || character
@@ -359,26 +393,12 @@ class Money
         else
           ""
         end
-      elsif rules[:html]
+      elsif rules[:html] || rules[:html_wrap]
         currency.html_entity == '' ? currency.symbol : currency.html_entity
       elsif rules[:disambiguate] && currency.disambiguate_symbol
         currency.disambiguate_symbol
       else
         money.symbol
-      end
-    end
-
-    def symbol_position_from(rules)
-      if rules.has_key?(:symbol_position)
-        if [:before, :after].include?(rules[:symbol_position])
-          return rules[:symbol_position]
-        else
-          raise ArgumentError, ":symbol_position must be ':before' or ':after'"
-        end
-      elsif currency.symbol_first?
-        :before
-      else
-        :after
       end
     end
   end